+++ packages/cli/api-importers/openapi/openapi-ir-parser/src/openapi/v3/converters/convertSecurityScheme.ts
@@ -39,6 +39,7 @@ function convertSecuritySchemeHelper(
     taskContext: TaskContext
 ): SecurityScheme | undefined {
     try {
+        const websocketAuthFallback = parseWebSocketAuthFallback(securityScheme);
         if (securityScheme.type === "apiKey" && securityScheme.in === "header") {
             const bearerFormat = getExtension<string>(securityScheme, OpenAPIExtension.BEARER_FORMAT);
             const headerNames = getExtension<HeaderSecuritySchemeNames>(
@@ -51,7 +52,8 @@ function convertSecuritySchemeHelper(
                 headerVariableName:
                     headerNames?.name ??
                     getExtension<string>(securityScheme, FernOpenAPIExtension.HEADER_VARIABLE_NAME),
-                headerEnvVar: headerNames?.env
+                headerEnvVar: headerNames?.env,
+                websocketAuthFallback
             });
         } else if (securityScheme.type === "http" && securityScheme.scheme?.toLowerCase() === "bearer") {
             // ^ case insensitivity for securityScheme.scheme required in OAS
@@ -59,7 +61,6 @@ function convertSecuritySchemeHelper(
                 securityScheme,
                 FernOpenAPIExtension.FERN_BEARER_TOKEN
             );
-            const websocketAuthFallback = parseWebSocketAuthFallback(securityScheme);
             return SecurityScheme.bearer({
                 tokenVariableName:
                     bearerNames?.name ??
@@ -77,17 +78,19 @@ function convertSecuritySchemeHelper(
                 usernameEnvVar: basicSecuritySchemeNamingAndEnvvar?.username?.env,
                 passwordVariableName:
                     basicSecuritySchemeNamingAndEnvvar?.password?.name ?? basicSecuritySchemeNaming.passwordVariable,
-                passwordEnvVar: basicSecuritySchemeNamingAndEnvvar?.password?.env
+                passwordEnvVar: basicSecuritySchemeNamingAndEnvvar?.password?.env,
+                websocketAuthFallback
             });
         } else if (securityScheme.type === "openIdConnect") {
             return SecurityScheme.bearer({
                 tokenVariableName: undefined,
                 tokenEnvVar: undefined,
-                websocketAuthFallback: undefined
+                websocketAuthFallback
             });
         } else if (securityScheme.type === "oauth2") {
             return SecurityScheme.oauth({
-                scopesEnum: getScopes(securityScheme, source)
+                scopesEnum: getScopes(securityScheme, source),
+                websocketAuthFallback
             });
         }
     } catch (error) {