+++ packages/cli/api-importers/openapi/openapi-ir/src/sdk/serialization/resources/commons/types/OauthSecurityScheme.ts
@@ -4,16 +4,19 @@ import type * as FernOpenapiIr from "../../../../api/index.js";
 import * as core from "../../../../core/index.js";
 import type * as serializers from "../../../index.js";
 import { EnumSchema } from "../../finalIr/types/EnumSchema.js";
+import { BaseSecurityScheme } from "./BaseSecurityScheme.js";
 
 export const OauthSecurityScheme: core.serialization.ObjectSchema<
     serializers.OauthSecurityScheme.Raw,
     FernOpenapiIr.OauthSecurityScheme
-> = core.serialization.objectWithoutOptionalProperties({
-    scopesEnum: EnumSchema.optional(),
-});
+> = core.serialization
+    .objectWithoutOptionalProperties({
+        scopesEnum: EnumSchema.optional(),
+    })
+    .extend(BaseSecurityScheme);
 
 export declare namespace OauthSecurityScheme {
-    export interface Raw {
+    export interface Raw extends BaseSecurityScheme.Raw {
         scopesEnum?: EnumSchema.Raw | null;
     }
 }